requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
pymongo>=4.6.1
argparse>=1.4.0
schedule>=1.2.0
paho-mqtt
pytz